requirements.txt
anthropic>=0.35.0
requests>=2.31.0
PyYAML>=6.0